Move LoginController into Auth namespace
